EventItem refactor and move
